requirements.txt
telethon==1.34.0
python-dotenv==1.0.1
asyncio==3.4.3
